Rollback one unintentional rename + Explain remaining TODOs as discussed.
