+++ core-rust/jni-export/src/lib.rs
@@ -81,7 +81,7 @@ fn export_extern_functions() {
     state_manager::jni::mempool::export_extern_functions();
     state_manager::jni::node_rust_environment::export_extern_functions();
     state_manager::jni::protocol_update::export_extern_functions();
-    state_manager::jni::system_executor::export_extern_functions();
+    state_manager::jni::state_computer::export_extern_functions();
     state_manager::jni::state_reader::export_extern_functions();
     state_manager::jni::transaction_preparer::export_extern_functions();
     state_manager::jni::transaction_store::export_extern_functions();
