+++ core-rust/state-manager/src/jni/mod.rs
@@ -68,8 +68,8 @@ pub mod mempool;
 pub mod node_rust_environment;
 pub mod prometheus;
 pub mod protocol_update;
+pub mod state_computer;
 pub mod state_reader;
-pub mod system_executor;
 pub mod test_state_reader;
 pub mod transaction_preparer;
 pub mod transaction_store;
