+++ core-rust/state-manager/src/jni/state_computer.rs
