+++ core-rust/state-manager/src/protocol/protocol_updates/definitions/mod.rs
@@ -48,7 +48,6 @@ impl UpdateBatchGenerator for ScryptoEntriesBatchGenerator {
                     })
                     .collect(),
             )),
-            // TODO(wip): return scenarios as consecutive batches
             _ => None,
         }
     }
